automated-data-governance-csv-simple/requirements.txt
streamlit
pandas
pyyaml
pyarrow
